Разбивает приложение на модули
