+++ js/data.js
@@ -1,0 +1,9 @@
+const CHECKIN_TIMES = ['12:00', '13:00', '14:00'];
+const CHECKOUT_TIMES = ['12:00', '13:00', '14:00'];
+const PHOTOS = ['https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg', 'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg', 'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg'];
+const TYPES = ['palace', 'flat', 'house', 'bungalow', 'hotel'];
+const FEATURES = ['wifi', 'dishwasher', 'parking', 'washer', 'elevator', 'conditioner'];
+const TITLES = ['Title 1','Title 2', 'Title 3', 'Title 4', 'Title 5', 'Title 6', 'Title 7', 'Title 8'];
+const DESCRIPTIONS = ['Descr 1','Descr 2', 'Descr 3', 'Descr 4', 'Descr 5'];
+
+export {CHECKIN_TIMES, CHECKOUT_TIMES, PHOTOS, TYPES, FEATURES, TITLES, DESCRIPTIONS};
